Rework roomIO
Reworks all comments in the roomIO package.
