+++ src/main/java/nl/tudelft/contextproject/roomIO/EntityReader.java
@@ -15,6 +15,7 @@ import nl.tudelft.contextproject.model.entities.Key;
 import nl.tudelft.contextproject.model.entities.PlayerTrigger;
 import nl.tudelft.contextproject.model.TickListener;
 import nl.tudelft.contextproject.model.entities.WallFrame;
+import nl.tudelft.contextproject.util.ReaderUtil;
 import nl.tudelft.contextproject.util.ScriptLoader;
 import nl.tudelft.contextproject.util.ScriptLoaderException;
 
@@ -22,21 +23,35 @@ import nl.tudelft.contextproject.util.ScriptLoaderException;
  * Utility class to read entities from a buffetredReader.
  */
 public final class EntityReader {
+
+	/**
+	 * Private constructor to avoid instantiation.
+	 */
 	private EntityReader() {}
 
 	/**
 	 * Read the specified amount of entities and add them to the set of entities.
-	 * @param entities The set to add all loaded entities to.
-	 * @param entityCount The number of entities to load.
-	 * @param xOffset The horizontal offset used for all entities.
-	 * @param yOffset The vertical offset used for all entities.
-	 * @param br The bufferedReader to read from.
-	 * @param path The path of the room folder.
-	 * @throws IOException When reading from the reader goes wrong.
-	 * @throws ScriptLoaderException When loading of a script goes wrong.
+	 *
+	 * @param entities
+	 * 		the set to add all loaded entities to
+	 * @param entityCount
+	 * 		the number of entities to load
+	 * @param xOffset
+	 * 		the horizontal offset used for all entities
+	 * @param yOffset
+	 * 		the vertical offset used for all entities
+	 * @param br
+	 * 		the bufferedReader to read from
+	 * @param path
+	 * 		the path of the room folder
+	 * @throws IOException
+	 * 		when reading from the reader goes wrong
+	 * @throws ScriptLoaderException
+	 * 		when loading of a script goes wrong
 	 */
 	public static void readEntities(Set<Entity> entities, int entityCount, int xOffset, int yOffset, 
 			BufferedReader br, String path) throws IOException, ScriptLoaderException {
+
 		for (int i = 0; i < entityCount; i++) {
 			String in = br.readLine();
 			if (in == null) throw new IllegalArgumentException("Empty line where some data was expected when loading entity[" + i + "].");
@@ -54,42 +69,53 @@ public final class EntityReader {
 
 	/**
 	 * Create an entity with specified type and location.
-	 * @param type The type of the entity to create.
-	 * @param x The x location of the entity.
-	 * @param y The y location of the entity.
-	 * @param z The z location of the entity.
-	 * @param data The full data array from the file, can be used to get additional information.
-	 * @param path The path of the file folder. 
-	 * @return The created entity.
-	 * @throws ScriptLoaderException When the loading of scripts goes wrong.
+	 *
+	 * @param type
+	 * 		the type of the entity to create
+	 * @param x
+	 * 		the x location of the entity
+	 * @param y
+	 * 		the y location of the entity
+	 * @param z
+	 * 		the z location of the entity
+	 * @param data
+	 * 		the full data array from the file, can be used to get additional information
+	 * @param path
+	 * 		the path of the file folder
+	 * @return
+	 * 		the created entity
+	 * @throws ScriptLoaderException
+	 * 		when the loading of scripts goes wrong
 	 */
 	protected static Entity getEntity(String type, float x, float y, float z, String[] data, String path) throws ScriptLoaderException {
 		ScriptLoader sl = new ScriptLoader(EntityReader.class.getResource(path).getPath());
 		Entity e;
 		if (data == null) throw new IllegalArgumentException("Data is null");
+
 		switch (type) {
-		case "Key":
-			if (data.length < 5) throw new IllegalArgumentException("Key must specify at least 5 values.");
-			e = new Key(LightReader.getColor(data[4]));
-			break;
-		case "Door":
-			if (data.length < 5) throw new IllegalArgumentException("Door must specify at least 5 values.");
-			e = new Door(LightReader.getColor(data[4]));
-			break;
-		case "Bomb":
-			e = new Bomb();
-			break;
-		case "PlayerTrigger":
-			if (data.length < 7) throw new IllegalArgumentException("PlayerTrigger must specify at least 7 values.");
-			e = new PlayerTrigger(Float.parseFloat(data[4]), Float.parseFloat(data[5]), 
-					sl.getInstanceOf(data[6], TickListener.class), new Vector3f(x, y, z));
-			return e;
-		case "WallFrame":
-			e = new WallFrame(new Vector3f(x, y, z), path + data[5], Direction.valueOf(data[4]));
-			return e;
-		default:
-			throw new IllegalArgumentException(type + " is not a known Entity type!");
+			case "Key":
+				if (data.length < 5) throw new IllegalArgumentException("Key must specify at least 5 values.");
+				e = new Key(ReaderUtil.getColor(data[4]));
+				break;
+			case "Door":
+				if (data.length < 5) throw new IllegalArgumentException("Door must specify at least 5 values.");
+				e = new Door(ReaderUtil.getColor(data[4]));
+				break;
+			case "Bomb":
+				e = new Bomb();
+				break;
+			case "PlayerTrigger":
+				if (data.length < 7) throw new IllegalArgumentException("PlayerTrigger must specify at least 7 values.");
+				e = new PlayerTrigger(Float.parseFloat(data[4]), Float.parseFloat(data[5]),
+						sl.getInstanceOf(data[6], TickListener.class), new Vector3f(x, y, z));
+				return e;
+			case "WallFrame":
+				e = new WallFrame(new Vector3f(x, y, z), path + data[5], Direction.valueOf(data[4]));
+				return e;
+			default:
+				throw new IllegalArgumentException(type + " is not a known Entity type!");
 		}
+
 		e.move(x, y, z);
 		return e;
 	}
