+++ src/test/java/nl/tudelft/contextproject/roomIO/LightReaderTest.java
@@ -57,35 +57,6 @@ public class LightReaderTest {
 	}
 	
 	/**
-	 * Test if getting a random color gives us a color.
-	 */
-	@Test
-	public void testGetColorRandom() {
-		ColorRGBA res = LightReader.getColor("randomColor");
-		assertNotNull(res);
-	}
-	
-	/**
-	 * Test getting a simple color.
-	 */
-	@Test
-	public void testGetColor() {
-		ColorRGBA res = LightReader.getColor(".5/.3/.2/.1");
-		assertEquals(.5, res.getRed(), 1e-6);
-		assertEquals(.3, res.getGreen(), 1e-6);
-		assertEquals(.2, res.getBlue(), 1e-6);
-		assertEquals(.1, res.getAlpha(), 1e-6);
-	}
-	
-	/**
-	 * Test for getting a color with a missing argument.
-	 */
-	@Test (expected = IllegalArgumentException.class)
-	public void testGetColorToFewArguments() {
-		LightReader.getColor(".5/.3/.2");
-	}
-	
-	/**
 	 * Read an ambient light from a string.
 	 * @throws IOException Should not happen.
 	 */
